feat : ScoreManager 구현 및 점수 관리 시스템 독립
1. ScoreManager 구현 (System)
- SingletonBehaviour 상속으로 전역 접근성 확보 및 점수 관리 모듈 독립
- UI 업데이트 로직 캡슐화

2. 점수 산정 로직 이관
- GameManager에 내부 Score 관련 AddScore, AddBonusScore 산출 공식을 ScoreManager로 이관

Ref #21
