+++ Assets/Scripts/ScoreManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5e5012f1749404101922bbb90d0ec2bd
